Feat: Team KillScore Update

+++ Source/GuardiansOrders/Game/GOGameState.h
@@ -79,10 +79,10 @@ public:
 	TArray<AGOPlayerState*> BlueTeam;
 
 	UPROPERTY(ReplicatedUsing = OnRep_RedTeamScore)
-	float RedTeamScore = 0.f;
+	int32 RedTeamScore = 0.f;
 
 	UPROPERTY(ReplicatedUsing = OnRep_BlueTeamScore)
-	float BlueTeamScore = 0.f;
+	int32 BlueTeamScore = 0.f;
 
 	UFUNCTION()
 	void OnRep_RedTeamScore();
